feat(memory): extract memory/ module from db.rs (mika#1259) (#1463)
* plan(1446): memory/ extraction — 5 accreted (validated) + 10 surfaced; 15 total scope

* feat(memory): extract 15 memory/-domain methods from db.rs (mika#1259)

Foundation §6 memory/ module - 15 methods relocated:
- Core memory (3), People (4), Preferences (3), Events (2),
  Facts (2), KG bridges (1).

Methods stay impl Database (zero call-site churn) per parent AC3.
Wave 2 sub-issue #1446.

Pipeline-Exempt: code-only - mechanical impl-block relocation per groomed plan.


---------
