+++ crates/mika-agent/src/lib.rs
@@ -10,6 +10,7 @@ pub mod db;
 pub(crate) mod github_graphql;
 pub mod kg;
 pub mod mcp;
+pub mod memory;
 pub mod messaging;
 pub mod operational;
 pub mod panic_hook;
